+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @prosopo/provider
 
+## 4.0.0
+### Major Changes
+
+- 41c54b0: Redis cache to avoid hitting the database multiple times per API request
+
+### Minor Changes
+
+- f14d645: Fix to catch NaN being passed in as score in sessions
+- 74092d0: Stream data back to central for decisions
+
+### Patch Changes
+
+- Updated dependencies [74092d0]
+  - @prosopo/types-database@4.6.0
+  - @prosopo/database@3.10.0
+  - @prosopo/types-env@2.9.4
+  - @prosopo/env@3.4.4
+  - @prosopo/api-express-router@3.1.4
+
 ## 3.24.0
 ### Minor Changes